Implementing the loan design patterns with Java Lambdas
This is a case where a static function takes in a lambda function. The static function creates an instance of an object, and the `loans` out that instance to the lambda to temporarily allow it to do what it wants with it.

Notice that the lambda doesn't need to return the provided instance since the function calling the lambda has that instance within its scope and so it automatically, `takes it back` when the lambda is done.